+++ Virtusize/Sources/VirtusizeConfiguration.swift
@@ -25,7 +25,7 @@
 import Foundation
 
 struct VirtusizeConfiguration {
-	static let SDKVersion = "2.12.5"
+	static let SDKVersion = "2.12.6"
     static let defaultAoyamaVersion = "3.4.2"
     static let resourceBundleName = "VirtusizeCore"
 }