+++ CHANGES.md
@@ -10,6 +10,10 @@ Use list notation, and following prefixes:
 - Bugfix - when fixing any major bug
 - Docs - for any improvement to documentation
 
+### 2.12.6
+- Feature: Add event listener to handle language change on web in Virtusize widgets
+- Fix: Display error state of VirtusizeInPageStandard widget
+
 ### 2.12.5
 - Fix: Compilation error when packages added via Swift Package Manager, due to missing explicit `import Foundation`